Move fanhelper to util package to help eliminate import cycles
